requirements.txt
# ========================================
# TELEGRAM BOT - ЗАЛЕЖНОСТІ ПРОЕКТУ
# ========================================
# Дата оновлення: 3 липня 2025
# Python версія: 3.12.3
# Операційна система: Ubuntu 24.04.2 LTS

# ========================================
# ОСНОВНІ ЗАЛЕЖНОСТІ БОТА
# ========================================

# Telegram Bot API - основна бібліотека для роботи з Telegram
python-telegram-bot>=22.1

# HTTP клієнти для API запитів  
httpx>=0.28.1
aiohttp>=3.12.10

# Конфігурація та змінні оточення
python-dotenv>=1.1.0

# Робота з YAML файлами (конфігурація полів)
pyyaml>=6.0.2

# ========================================
# ІНТЕГРАЦІЇ З ЗОВНІШНІМИ СЕРВІСАМИ
# ========================================

# Google Sheets інтеграція
gspread>=6.2.1
google-auth>=2.40.3
google-auth-oauthlib>=1.2.2
oauth2client>=4.1.3

# ========================================
# СИСТЕМНІ ЗАЛЕЖНОСТІ
# ========================================

# HTTP бібліотеки
requests>=2.32.3
httplib2>=0.22.0

# Асинхронність
aiohappyeyeballs>=2.6.1
aiosignal>=1.3.2
anyio>=4.9.0

# Криптографія та безпека
certifi>=2025.4.26
rsa>=4.9.1

# Парсинг та валідація
attrs>=25.3.0
charset-normalizer>=3.4.2
idna>=3.10
pyparsing>=3.2.3

# Типізація
typing_extensions>=4.14.0

# ========================================
# ДОДАТКОВІ УТИЛІТИ  
# ========================================

# Кешування
cachetools>=5.5.2

# Робота з URL та багатозначними словниками
yarl>=1.20.0
multidict>=6.4.4
frozenlist>=1.6.2
propcache>=0.3.1

# OAuth аутентифікація
oauthlib>=3.2.2
requests-oauthlib>=2.0.0

# ASN.1 підтримка для Google API
pyasn1>=0.6.1
pyasn1_modules>=0.4.2

# HTTP/1.1 підтримка
h11>=0.16.0
httpcore>=1.0.9

# Визначення мережі
sniffio>=1.3.1

# Утиліти
six>=1.17.0
urllib3>=2.4.0